Random maps now render all the rooms and corridors correctly. Added new sprites everywhere.

+++ engine/components/animation_component.h
@@ -13,11 +13,11 @@ class animation_component : public component {
   void addSprite(sprite_component s);
   void addSprite(sprite_id spriteId, rect src, int width, int height);
   void setAnimationTicks(int t);
-  void setScale(int scale);
+  void setScale(float scale);
   void setLooping(bool l);
 
   sprite_component tickAndGetSprite();
-  int getScale();
+  float getScale();
 
  private:
   std::vector<sprite_component> _sprites;
@@ -25,7 +25,7 @@ class animation_component : public component {
   int _currentSprite = 0;
   int _animationTicks = 10;
   int _currentTicks = 0;
-  int _scale = 1;
+  float _scale = 1.0;
 };
 
 }  // namespace components
